requirements.txt
python-slugify==1.2.5
Unidecode==1.0.22
